refactor: move the range gadgets into composer::range
Pure code motion: `component_range_bits`, `component_range`,
`range_check` and `range_check_even` move out of `composer.rs` into a
`range` submodule, together with their soundness tests. `range_check`
becomes `pub(super)` so the callers left behind can still reach it.
